+++ lib/generators/administrate/install/templates/routes.rb
@@ -1,7 +1,0 @@
-namespace :admin do
-    DashboardManifest::DASHBOARDS.each do |dashboard_resource|
-      resources dashboard_resource
-    end
-
-    root controller: DashboardManifest::ROOT_DASHBOARD, action: :index
-  end