+++ spec/lib/administrate/search_spec.rb
@@ -1,5 +1,8 @@
 require "spec_helper"
 require "support/constant_helpers"
+require "administrate/field/string"
+require "administrate/field/email"
+require "administrate/field/number"
 require "administrate/search"
 
 # NOTE: Dashboard mocks (MockDashboard, DashboardWithAnArrayOfScopes and